requirements.txt
PyQt5==5.15.10
requests==2.31.0
matplotlib==3.8.2
python-dotenv==1.0.1